// infrastructure/lib/common/enum.ts
export enum Stage {
  DESA = 'DESA',
  TEST = 'TEST',
  PROD = 'PROD',
}

export enum EncodingType {
  UTF8 = 'utf-8',
  UTF16 = 'utf-16',
  UTF32 = 'utf-32',
  ASCII = 'ascii',
  BASE64 = 'base64',
  HEX = 'hex',
}

export enum HashAlgorithm {
  SHA256 = 'sha256',
  MD5 = 'md5',
}

export enum Region {
  US_EAST_1 = 'us-east-1',
}

export enum RegionAbrev {
  UE1 = 'UE1',
}

export enum AccountAbrev {
  DEVL = 'DEVL',
  TEST = 'TEST',
  PROD = 'PROD',
}

export enum TAG {
  NAME = 'Name',
  ENTORNO = 'Entorno',
  AMBIENTE = 'Ambiente',
  PROYECTO = 'Proyecto',
  RESPONSABLE = 'Responsable',
}